ocr_service/requirements.txt
# OCR 服务依赖

# Web 框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# OCR 引擎
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# 图像处理
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# HTTP 客户端（用于健康检查等）
httpx>=0.24.0